requirements.txt
openai-whisper
Flask